+++ up-gutenberg-query-filter.php
@@ -4,8 +4,8 @@
  * Description:       Filter blocks for the query loop utilising the interactivity API.
  * Requires at least: 6.6
  * Requires PHP:      8.0
- * Version:           1.1.6
- * Author:            Upcoder
+ * Version:           2.1.0
+ * Author:            GEHIN Nicolas
  * License:           GPL-2.0-or-later
  * License URI:       https://www.gnu.org/licenses/gpl-2.0.html
  * Text Domain:       up-gutenberg-query-filter